+++ common/src/main/java/koala/fishingreal/FishingReal.java
@@ -35,8 +35,8 @@ public class FishingReal {
     public static Entity convertItemStack(ItemStack itemstack, Player player) {
         if (player != null && player.level instanceof ServerLevel serverLevel) {
             FishingConversion.FishingResult result = FishingReal.FISHING_MANAGER.getConversionResultFromStack(itemstack);
-            Entity resultEntity = result.entity().create(player.level);
-            if (resultEntity != null) {
+            if(result != null) {
+                Entity resultEntity = result.entity().create(player.level);
                 result.tag().ifPresent(resultEntity::load);
                 if (result.randomizeNbt() && resultEntity instanceof Mob resultMob) {
                     resultMob.finalizeSpawn(serverLevel, serverLevel.getCurrentDifficultyAt(player.blockPosition()), MobSpawnType.NATURAL, null, null);
@@ -57,24 +57,5 @@ public class FishingReal {
             }
         }
         return fishedEntity;
-    }
-
-    /**
-     * Rip of FishingHook#Retrieve logic for use where direct entity conversion can't occur (Forge API)
-     */
-    public static void fishUpEntity(Entity entity, FishingHook hook, ItemStack stack, Player player) {
-        double dX = player.getX() - hook.getX();
-        double dY = player.getY() - hook.getY();
-        double dZ = player.getZ() - hook.getZ();
-        double m = 0.12;
-        entity.setDeltaMovement(dX * m, dY * m + Math.sqrt(Math.sqrt(dX * dX + dY * dY + dZ * dZ)) * 0.08, dZ * m);
-        player.level.addFreshEntity(entity);
-        player.level.addFreshEntity(new ExperienceOrb(player.level, player.getX(), player.getY() + 0.5, player.getZ() + 0.5, player.level.random.nextInt(6) + 1));
-        if (stack.is(ItemTags.FISHES)) {
-            player.awardStat(Stats.FISH_CAUGHT, 1);
-        }
-        if (player instanceof ServerPlayer serverPlayer) {
-            CriteriaTriggers.FISHING_ROD_HOOKED.trigger(serverPlayer, player.getUseItem(), hook, List.of(stack));
-        }
     }
 }