+++ fabric/src/main/java/koala/fishingreal/fabric/mixin/FishingHookMixin.java
@@ -1,4 +1,4 @@
-package koala.fishingreal.mixin;
+package koala.fishingreal.fabric.mixin;
 
 import koala.fishingreal.FishingReal;
 import net.minecraft.world.entity.Entity;